demo/requirements.txt
# Core dependencies
pdfplumber==0.10.3
python-docx==0.8.11
PyPDF2==3.0.1

# NLP & Extraction
spacy==3.7.2
python-dotenv==1.0.0

# Embeddings & Vector search
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.0
scikit-learn==1.3.2

# API & Web
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Dashboard
streamlit==1.28.1
plotly==5.18.0

# Data handling
pandas==2.1.3
openpyxl==3.11.0

# Async & Background jobs
aiofiles==23.2.1

# Utils
python-dateutil==2.8.2
requests==2.31.0

# Testing & Logging
python-logging-loki==0.3.2
